requirements.txt
absl-py==0.7.1
appdirs==1.4.3
astor==0.8.0
attrs==19.1.0
backports.weakref==1.0.post1
black==18.9b0
Click==7.0
enum34==1.1.6
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.17.0
protobuf==3.9.1
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-hub==0.5.0
termcolor==1.1.0
toml==0.10.0
tqdm==4.32.2
Werkzeug==0.15.5
wrapt==1.11.2
